Fix incorrectly displayed dropdown values.
I'm not sure if this _really_ is necessary, but it does seem to be.  If
the user manually chose lucene (they selected cloudsearch and then
switched to lucene while on the title field), and then selected the
timestamp field (forcing it to be cloudsearch), and then scroll back to
a non-timestamp field, the dropdown will still display 'cloudsearch',
but a lucene query will be generated.

The fix (which I doubt is the best way of doing this) is to always set
the dropdown value to the used syntax.